refactor: NoticeDetailInfoCard로 컴포넌트 변경 및 리팩토링
