+++ src/constants/applicationStatus.ts
@@ -1,0 +1,6 @@
+export const APPLICATION_STATUS = {
+  PENDING: "pending",
+  ACCEPTED: "accepted",
+  REJECTED: "rejected",
+  CANCELED: "canceled",
+} as const;
